electron4: make DIST option ON by default and include distribution zip files in package.
electron (for linux, mac, and windows) is distributed as a zip file at
GitHub releases. I provide similar zip files at my GitHub repo (for
use with electron-download).

FreeBSD packages are often required to be rebuilt due to version bumps of
dependencies. Every time electron package is updated, it is necessary
to rebuild the zip files as well. It is possible but burdensome.

So I decided to simply include the zip files as part of electron
package. Those zip files can be used with electron-download by copying
them into a pre-defined cache directory.