x11/hyper: add port: Terminal built on web technologies
The goal of the project is to create a beautiful and extensible
experience for command-line interface users, built on open web
standards. In the beginning, our focus will be primarily around speed,
stability and the development of the correct API for extension
authors.

In the future, we anticipate the community will come up with
innovative additions to enhance what could be the simplest, most
powerful and well-tested interface for productivity.

WWW: https://hyper.is