deskutils/notesnook: add port: Fully open source & end-to-end encrypted note taking application
Notesnook is a free (as in speech) & open-source note-taking app
focused on user privacy & ease of use. To ensure zero knowledge
principles, Notesnook encrypts everything on your device using
XChaCha20-Poly1305 & Argon2.

Notesnook is our proof that privacy does not (always) have to come at
the cost of convenience. We aim to provide users peace of mind & 100%
confidence that their notes are safe and secure. The decision to go
fully open source is one of the most crucial steps towards that.

WWW: https://notesnook.com/

This port is not (yet) compatible with poudriere.