deskutils/joplin-desktop: update to 3.3.12 (WIP)
This is still work in progress. The port builds fine but currently it
fails to start with the error "TypeError: LRU is not a
constructor". Something appears wrong with the versions of semver and
lru-cache used under the hood. I still cannot figure out why the error
is happening.

Changelog: https://joplinapp.org/news/20250428-release-3-3