devel/electron37: unbreak non-root build with poudriere
The source archive of chromium hosted at GitHub is splitted into
multiple fragments due to size constraints. So we have to concatenate
them into a single tarball before extracting it.

Previously, the combined tarball was written to DISTDIR/DIST_SUBDIR,
which may not be writable by the poudriere user. So change the
directory to WRKSRC:H and manually extracting the tarball to avoid
such breakage.