electron.mk: Attempt hierarchical module install only when guiding PKGJSONSDIR exists
When a project contains multiple lockfiles, it usually means we need
to do "npm install" in each directory containing respective
lockfile.

When we have pre-stored package.jsons and lockfiles in PKGJSONSDIR, we
can utilize those files as guiding information. However, there may be
the case we don't have PKGJSONSDIR. In that case, we assume "npm
install" is only needed in the top-level directory of the project.

This might be incomplete but I think it is a reasonable assumtion for
now.