electron.mk: Change directory for electron binaries to under WRKDIR
Sometimes, the build process requires electron binary to be
writable (for security fusing purposes). The electron executable under
LOCALBASE is not writable (perm 555).

We copy electron distribution files from the directory under LOCALBASE
to the one under WRKDIR and change permissions to 555 to 755 before
the pre-build phase. We should utilize the directory for packaging
electron applications as well.